# src/swagger-output.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "CompassCar API",
		"description": "API para gerenciamento de carros, incluindo funcionalidades de criação, leitura, atualização e exclusão de registros de carros. Esta API permite a integração com outros sistemas, fornecendo endpoints seguros e documentados para operações de CRUD. Aproveite as funcionalidades avançadas para gerenciar seu inventário de carros de forma eficiente.",
		"version": "1.0.0"
	},
	"host": "localhost:3000",
	"basePath": "/",
...
	"schemes": ["http"],
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"components": {
		"securitySchemes": {
			"BearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Token de Autenticação"
			}
		}
	},
	"security": [
		{
			"BearerAuth": []
		}
	],
	"paths": {
...
		"/api/v1/rental-order": {
			"post": {
				"tags": ["Order"],
				"summary": "Criar um novo pedido",
				"description": "Essa rota permite a criação de um novo pedido no sistema.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"car_id": {
										"type": "string",
										"example": "3e2846a7-8d26-46ad-80ee-82029724a761",
										"description": "ID do Carro"
									},
									"customer_id": {
										"type": "string",
										"example": "a8052f5b-abdc-48b0-aa52-e8eb64a5cd76",
										"description": "ID do Cliente"
									}
								},
								"required": ["car_id", "customer_id"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Pedido criado com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"example": "e57e3a2d-0a16-4f8f-99be-64723bc1d396"
										}
									}
								}
							}
						}
					},
					"404_CustomerNotFound": {
						"description": "Cliente não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Customer not found"
										}
									}
								}
							}
						}
					},
					"404_CarNotFound": {
						"description": "Carro não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Car not found"
										}
									}
								}
							}
						}
					},
					"400_CarInUse": {
						"description": "Carro já está em uso",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Car can't be used. It is already in use"
										}
									}
								}
							}
						}
					},
					"400_CustomerOrderOpen": {
						"description": "Cliente já possui uma ordem em aberto",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Can't create order because customer has an opened order already"
										}
									}
								}
							}
						}
					},
					"409": {
						"description": "Carro já registrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "A car with this plate already exists"
										}
									}
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": ["Order"],
				"summary": "Listar todos os pedidos",
				"description": "Essa rota permite a listagem de todos os pedidos no sistema.",
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"example": 1
						},
						"description": "Número da página para paginação"
					},
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"example": 10
						},
						"description": "Número de itens por página"
					},
					{
						"name": "status",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "ativo | inativo | excluido"
						},
						"description": "Filtrar por status do carro"
					},
					{
						"name": "plate",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "MKF3123"
						},
						"description": "Filtra pela placa do carro"
					},
					{
						"name": "brand",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "Ford"
						},
						"description": "Filtra pela marca do carro"
					},
					{
						"name": "model",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "Focus"
						},
						"description": "Filtra pelo modelo do carro"
					},
					{
						"name": "km",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "10.800"
						},
						"description": "Filtra pelo total em Kms do carro"
					},
					{
						"name": "fromYear",
						"in": "query",
						"required": false,
						"schema": {
							"type": "number",
							"example": 2014
						},
						"description": "Filtra os carros fabricados a partir do ano especificado"
					},
					{
						"name": "untilYear",
						"in": "query",
						"required": false,
						"schema": {
							"type": "number",
							"example": 2020
						},
						"description": "Filtra os carros fabricados até o ano especificado"
					},
					{
						"name": "minDailyPrice",
						"in": "query",
						"required": false,
						"schema": {
							"type": "number",
							"example": 25000
						},
						"description": "Filtra os carros com preço diário a partir do valor especificado"
					},
					{
						"name": "maxDailyPrice",
						"in": "query",
						"required": false,
						"schema": {
							"type": "number",
							"example": 40000
						},
						"description": "Filtra os carros com preço diário até o valor especificado"
					},
					{
						"name": "order",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "ASC | DESC"
						},
						"description": "Ordem de classificação (ASC ou DESC)"
					},
					{
						"name": "sortBy",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"example": "year, km, daily_price"
						},
						"description": "Campo pelo qual a listagem será ordenada (year, km, daily_price)"
					}
				],
				"responses": {
					"200": {
						"description": "Lista de carros retornada com sucesso",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"count": {
											"type": "integer",
											"example": 4
										},
										"total_pages": {
											"type": "integer",
											"example": 1
										},
										"total": {
											"type": "integer",
											"example": 100
										},
										"current_page": {
											"type": "integer",
											"example": 1
										},
										"cars": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "string",
														"example": "e6d7f7d3-4e32-42e3-837a-86753bcae83b"
													},
													"plate": {
														"type": "string",
														"example": "MAS1453"
													},
													"brand": {
														"type": "string",
														"example": "Ford"
													},
													"model": {
														"type": "string",
														"example": "Focus"
													},
													"year": {
														"type": "integer",
														"example": 2018
													},
													"daily_price": {
														"type": "number",
														"example": 16.8
													},
													"km": {
														"type": "number",
														"example": 18000.0
													},
													"status": {
														"type": "string",
														"enum": ["ativo", "inativo", "excluído"],
														"example": "inativo"
													},
													"created_at": {
														"type": "string",
														"example": "2024-10-14T15:53:52.227Z"
													},
													"updated_at": {
														"type": "string",
														"example": "2024-10-14T16:50:17.119Z"
													},
													"items": {
														"type": "array",
														"items": {
															"type": "object",
															"properties": {
																"id": {
																	"type": "string",
																	"example": "fb901f0e-842b-44ca-8b87-f3309c94fe71"
																},
																"item": {
																	"type": "string",
																	"example": "vidro eletrico"
																}
															}
														}
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Nenhum carro encontrado.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Cars not found"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/rental-order/{id}": {
			"get": {
				"tags": ["Order"],
				"summary": "Busca um pedido por ID",
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"example": "e6d7f7d3-4e32-42e3-837a-86753bcae83b",
											"description": "Identificador único do carro"
										},
										"plate": {
											"type": "string",
											"example": "MKF3123",
											"description": "Placa do carro"
										},
										"brand": {
											"type": "string",
											"example": "Ford",
											"description": "Marca do carro"
										},
										"model": {
											"type": "string",
											"example": "Fusion",
											"description": "Modelo do carro"
										},
										"year": {
											"type": "number",
											"example": 2019,
											"description": "Ano de fabricação do carro"
										},
										"daily_price": {
											"type": "number",
											"example": 16.8,
											"description": "Diária"
										},
										"km": {
											"type": "number",
											"example": 30000,
											"description": "Quilometragem do carro"
										},
										"status": {
											"type": "string",
											"enum": ["ativo", "inativo", "excluido"],
											"example": "ativo",
											"description": "Status do carro"
										},
										"created_at": {
											"type": "string",
											"example": "2024-10-12T00:48:11.146Z",
											"description": "Data de criação do registro"
										},
										"updated_at": {
											"type": "string",
											"example": "2024-10-12T00:48:11.146Z",
											"description": "Data da última atualização do registro"
										},
										"items": {
											"type": "array",
											"items": {
												"type": "object",
												"properties": {
													"id": {
														"type": "string",
														"example": "89c00391-2d47-496c-8ca6-d560f3ed00a5",
														"description": "Identificador único do item"
													},
													"item": {
														"type": "string",
														"example": "vidro eletrico",
														"description": "Nome do item"
													}
												}
											},
											"description": "Lista de itens associados ao carro"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Carro não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Car not found"
										}
									}
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Order"],
				"description": "",
				"summary": "Atualiza um usuário por ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"fullName": {
										"type": "string",
										"example": "Teste da Silva",
										"description": "Nome completo do usuário"
									},
									"email": {
										"type": "string",
										"example": "teste@mail.com",
										"description": "E-mail do usuário"
									},
									"password": {
										"type": "string",
										"example": "teste123",
										"description": "Senha do usuário"
									}
								}
							},
							"required": []
						}
					}
				},
				"responses": {
					"200": {
						"description": "Usuário atualizado com sucesso. Se o corpo da requisição estiver vazio, nenhum dado será alterado.",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"id": {
											"type": "string",
											"example": "30503c46-69da-424f-857a-d745a68ed4e0"
										},
										"createdAt": {
											"type": "string",
											"example": "2024-10-11T19:55:11.499Z"
										},
										"deletedAt": {
											"type": "string",
											"example": null
										}
									}
								}
							}
						}
					},
					"400": {
						"description": "Requisição inválida",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Dados inválidos"
										}
									}
								}
							}
						}
					},
					"404": {
						"description": "Usuário não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "User not found"
										}
									}
								}
							}
						}
					}
				}
			},
			"delete": {
				"tags": ["Order"],
				"description": "",
				"summary": "Deleta um pedido por ID",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"204": {
						"description": "Pedido deletado com sucesso. Nenhum conteúdo retornado."
					},
					"404": {
						"description": "Pedido não encontrado",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"message": {
											"type": "string",
											"example": "Order not found"
										}
									}
								}
							}
						}
					}
				}
			}
		}
	},
...
}